requirements.txt
asgiref==3.8.1
Brotli==1.1.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.4
cssselect2==0.8.0
defusedxml==0.7.1
Django==5.2.3
django-ckeditor==6.7.3
django-js-asset==3.1.2
fonttools==4.58.4
idna==3.10
oauthlib==3.3.1
pillow==11.2.1
pycparser==2.22
pydyf==0.11.0
PyJWT==2.10.1
pyphen==0.17.2
python-dotenv==1.1.1
python3-openid==3.2.0
requests==2.32.4
requests-oauthlib==2.0.0
social-auth-app-django==5.4.3
social-auth-core==4.6.1
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
urllib3==2.5.0
weasyprint==65.1
webencodings==0.5.1
zopfli==0.2.3.post1
